requirements.txt
numpy
pillow 
pytesseract